indexing_pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
dataclasses-json==0.6.7
faiss-cpu==1.10.0
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
html2text==2024.2.26
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==0.3.83
langchain-openai==0.3.7
langchain-text-splitters==1.1.0
langsmith==0.3.45
lxml==6.0.2
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
notebook_shim==0.2.4
numpy==2.4.2
openai==1.109.1
orjson==3.11.7
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==82.0.0
SQLAlchemy==2.0.46
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
wheel==0.46.3
yarl==1.22.0
zstandard==0.23.0
